Module clap::_derive::_cookbook::cargo_example_derive

Available on crate feature unstable-doc only.

Example: cargo subcommand (Derive API)

use clap::Parser;

#[derive(Parser)] // requires `derive` feature
#[command(name = "cargo")]
#[command(bin_name = "cargo")]
enum Cargo {
    ExampleDerive(ExampleDerive),
}

#[derive(clap::Args)]
#[command(author, version, about, long_about = None)]
struct ExampleDerive {
    #[arg(long)]
    manifest_path: Option<std::path::PathBuf>,
}

fn main() {
    let Cargo::ExampleDerive(args) = Cargo::parse();
    println!("{:?}", args.manifest_path);
}

For more on creating a custom subcommand, see the cargo book. The crate clap-cargo can help in mimicking cargo’s interface.

The help looks like:

$ cargo-example-derive --help
Usage: cargo <COMMAND>

Commands:
  example-derive  A simple to use, efficient, and full-featured Command Line Argument Parser
  help            Print this message or the help of the given subcommand(s)

Options:
  -h, --help  Print help information

$ cargo-example-derive example-derive --help
A simple to use, efficient, and full-featured Command Line Argument Parser

Usage: cargo example-derive [OPTIONS]

Options:
      --manifest-path <MANIFEST_PATH>  
  -h, --help                           Print help information
  -V, --version                        Print version information

Then to directly invoke the command, run:

$ cargo-example-derive example-derive
None

$ cargo-example-derive example-derive --manifest-path Cargo.toml
Some("Cargo.toml")