Module clap::app

Re-exports

pub use self::settings::AppSettings;

Modules

help

meta

parser

settings

usage

validator

Structs

App
Used to create a representation of a command line program and all possible command line arguments. Application settings are set using the "builder pattern" with the App::get_matches family of methods being the terminal methods that starts the runtime-parsing process. These methods then return information about the user supplied arguments (or lack there of).