Module clap::app::help

Modules

term_size

Structs

Help
clap Help Writer.

Enums

CopyUntilResult
Possible results for a copying function that stops when a given byte was found.

Constants

TAB

Traits

ArgWithDisplay

ArgWithOrder

Functions

as_arg_trait

copy_and_capture
Copies the contents of a reader into a writer until a {tag} is found, copying the tag content to a buffer and returning its size. In addition to errors, there are three possible outputs:

copy_until
Copies the contents of a reader into a writer until a delimiter byte is found. On success, the total number of bytes that were copied from reader to writer is returned.

str_width

wrap_help