Function clap::app::help::copy_until

fn copy_until<R: Read, W: Write>(
    r: &mut R, 
    w: &mut W, 
    delimiter_byte: u8
) -> CopyUntilResult

Copies the contents of a reader into a writer until a delimiter byte is found. On success, the total number of bytes that were copied from reader to writer is returned.