Struct clap::app::help::Help

pub struct Help<'a> {
    writer: &'a mut Write,
    next_line_help: bool,
    hide_pv: bool,
    term_w: usize,
    color: bool,
    cizer: Colorizer,
    longest: usize,
    force_next_line: bool,
    use_long: bool,
}

clap Help Writer.

Wraps a writer stream providing different methods to generate help for clap objects.

Fields
writer: &'a mut Write next_line_help: bool hide_pv: bool term_w: usize color: bool cizer: Colorizer longest: usize force_next_line: bool use_long: bool
Methods

impl<'a> Help<'a>

pub fn new(
w: &'a mut Write,
next_line_help: bool,
hide_pv: bool,
color: bool,
cizer: Colorizer,
term_w: Option<usize>,
max_w: Option<usize>,
use_long: bool
) -> Self

Create a new Help instance.

pub fn write_app_help(
w: &'a mut Write,
app: &App,
use_long: bool
) -> ClapResult<()>

Reads help settings from an App and write its help to the wrapped stream.

pub fn write_parser_help(
w: &'a mut Write,
parser: &Parser,
use_long: bool
) -> ClapResult<()>

Reads help settings from a Parser and write its help to the wrapped stream.

pub fn write_parser_help_to_stderr(
w: &'a mut Write,
parser: &Parser
) -> ClapResult<()>

Reads help settings from a Parser and write its help to the wrapped stream which will be stderr. This method prevents formatting when required.

pub fn _write_parser_help(
w: &'a mut Write,
parser: &Parser,
stderr: bool,
use_long: bool
) -> ClapResult<()>

pub fn write_help(&mut self, parser: &Parser) -> ClapResult<()>

Writes the parser help to the wrapped stream.

impl<'a> Help<'a>

fn write_args_unsorted<'b: 'd, 'c: 'd, 'd, I: 'd>(
&mut self,
args: I
) -> Result<()> where
I: Iterator<Item = &'d ArgWithOrder<'b, 'c>>,

Writes help for each argument in the order they were declared to the wrapped stream.

fn write_args<'b: 'd, 'c: 'd, 'd, I: 'd>(&mut self, args: I) -> Result<()> where
I: Iterator<Item = &'d ArgWithOrder<'b, 'c>>,

Sorts arguments by length and display order and write their help to the wrapped stream.

fn write_arg<'b, 'c>(&mut self, arg: &ArgWithDisplay<'b, 'c>) -> Result<()>

Writes help for an argument to the wrapped stream.

fn short<'b, 'c>(&mut self, arg: &ArgWithDisplay<'b, 'c>) -> Result<()>

Writes argument's short command to the wrapped stream.

fn long<'b, 'c>(&mut self, arg: &ArgWithDisplay<'b, 'c>) -> Result<()>

Writes argument's long command to the wrapped stream.

fn val<'b, 'c>(&mut self, arg: &ArgWithDisplay<'b, 'c>) -> Result<String, Error>

Writes argument's possible values to the wrapped stream.

fn write_before_after_help(&mut self, h: &str) -> Result<()>

fn help<'b, 'c>(
&mut self,
arg: &ArgWithDisplay<'b, 'c>,
spec_vals: &str
) -> Result<()>

Writes argument's help to the wrapped stream.

fn spec_vals(&self, a: &ArgWithDisplay) -> String

impl<'a> Help<'a>

pub fn write_all_args(&mut self, parser: &Parser) -> ClapResult<()>

Writes help for all arguments (options, flags, args, subcommands) including titles of a Parser Object to the wrapped stream.

fn write_subcommands(&mut self, parser: &Parser) -> Result<()>

Writes help for subcommands of a Parser Object to the wrapped stream.

fn write_version(&mut self, parser: &Parser) -> Result<()>

Writes version of a Parser Object to the wrapped stream.

fn write_bin_name(&mut self, parser: &Parser) -> Result<()>

Writes binary name of a Parser Object to the wrapped stream.

pub fn write_default_help(&mut self, parser: &Parser) -> ClapResult<()>

Writes default help for a Parser Object to the wrapped stream.

impl<'a> Help<'a>

fn write_templated_help(
&mut self,
parser: &Parser,
template: &str
) -> ClapResult<()>

Write help to stream for the parser in the format defined by the template.

Tags arg given inside curly brackets: Valid tags are: * {bin} - Binary name. * {version} - Version number. * {author} - Author information. * {usage} - Automatically generated or given usage string. * {all-args} - Help for all arguments (options, flags, positionals arguments, and subcommands) including titles. * {unified} - Unified help for options and flags. * {flags} - Help for flags. * {options} - Help for options. * {positionals} - Help for positionals arguments. * {subcommands} - Help for subcommands. * {after-help} - Info to be displayed after the help message. * {before-help} - Info to be displayed before the help message.

The template system is, on purpose, very simple. Therefore the tags have to writen in the lowercase and without spacing.