Struct clap::app::parser::Parser

pub struct Parser<'a, 'b> where
    'a: 'b,  {
    pub meta: AppMeta<'b>,
    settings: AppFlags,
    pub g_settings: AppFlags,
    pub flags: Vec<FlagBuilder<'a, 'b>>,
    pub opts: Vec<OptBuilder<'a, 'b>>,
    pub positionals: VecMap<PosBuilder<'a, 'b>>,
    pub subcommands: Vec<App<'a, 'b>>,
    pub groups: Vec<ArgGroup<'a>>,
    pub global_args: Vec<Arg<'a, 'b>>,
    pub required: Vec<&'a str>,
    pub r_ifs: Vec<(&'a str, &'b str, &'a str)>,
    pub overrides: Vec<(&'b str, &'a str)>,
    help_short: Option<char>,
    version_short: Option<char>,
    cache: Option<&'a str>,
    pub help_message: Option<&'a str>,
    pub version_message: Option<&'a str>,
}

Fields
meta: AppMeta<'b> settings: AppFlags g_settings: AppFlags flags: Vec<FlagBuilder<'a, 'b>> opts: Vec<OptBuilder<'a, 'b>> positionals: VecMap<PosBuilder<'a, 'b>> subcommands: Vec<App<'a, 'b>> groups: Vec<ArgGroup<'a>> global_args: Vec<Arg<'a, 'b>> required: Vec<&'a str> r_ifs: Vec<(&'a str, &'b str, &'a str)> overrides: Vec<(&'b str, &'a str)> help_short: Option<char> version_short: Option<char> cache: Option<&'a str> help_message: Option<&'a str> version_message: Option<&'a str>
Methods

impl<'a, 'b> Parser<'a, 'b> where
'a: 'b,

pub fn with_name(n: String) -> Self

pub fn help_short(&mut self, s: &str)

pub fn version_short(&mut self, s: &str)

pub fn gen_completions_to<W: Write>(&mut self, for_shell: Shell, buf: &mut W)

pub fn gen_completions(&mut self, for_shell: Shell, od: OsString)

fn app_debug_asserts(&mut self) -> bool

fn debug_asserts(&self, a: &Arg) -> bool

fn add_conditional_reqs(&mut self, a: &Arg<'a, 'b>)

fn add_arg_groups(&mut self, a: &Arg<'a, 'b>)

fn add_reqs(&mut self, a: &Arg<'a, 'b>)

fn implied_settings(&mut self, a: &Arg<'a, 'b>)

pub fn add_arg(&mut self, a: Arg<'a, 'b>)

pub fn add_arg_ref(&mut self, a: &Arg<'a, 'b>)

pub fn add_group(&mut self, group: ArgGroup<'a>)

pub fn add_subcommand(&mut self, subcmd: App<'a, 'b>)

pub fn propagate_settings(&mut self)

pub fn derive_display_order(&mut self)

pub fn required(&self) -> Iter<&str>

pub fn has_args(&self) -> bool

pub fn has_opts(&self) -> bool

pub fn has_flags(&self) -> bool

pub fn has_positionals(&self) -> bool

pub fn has_subcommands(&self) -> bool

pub fn has_visible_opts(&self) -> bool

pub fn has_visible_flags(&self) -> bool

pub fn has_visible_positionals(&self) -> bool

pub fn has_visible_subcommands(&self) -> bool

pub fn is_set(&self, s: AS) -> bool

pub fn set(&mut self, s: AS)

pub fn unset(&mut self, s: AS)

pub fn verify_positionals(&mut self) -> bool

pub fn propagate_globals(&mut self)

fn possible_subcommand(&self, arg_os: &OsStr) -> (bool, Option<&str>)

fn parse_help_subcommand<I, T>(&self, it: &mut I) -> ClapResult<ParseResult<'a>> where
I: Iterator<Item = T>,
T: Into<OsString>,

fn is_new_arg(&mut self, arg_os: &OsStr, needs_val_of: ParseResult) -> bool

pub fn get_matches_with<I, T>(
&mut self,
matcher: &mut ArgMatcher<'a>,
it: &mut Peekable<I>
) -> ClapResult<()> where
I: Iterator<Item = T>,
T: Into<OsString> + Clone,

fn remove_overrides(&mut self, matcher: &mut ArgMatcher)

fn propagate_help_version(&mut self)

fn build_bin_names(&mut self)

fn parse_subcommand<I, T>(
&mut self,
sc_name: &str,
matcher: &mut ArgMatcher<'a>,
it: &mut Peekable<I>
) -> ClapResult<()> where
I: Iterator<Item = T>,
T: Into<OsString> + Clone,

pub fn groups_for_arg(&self, name: &str) -> Option<Vec<&'a str>>

pub fn args_in_group(&self, group: &str) -> Vec<String>

pub fn arg_names_in_group(&self, group: &str) -> Vec<&'a str>

pub fn create_help_and_version(&mut self)

fn check_for_help_and_version_str(&self, arg: &OsStr) -> ClapResult<()>

fn check_for_help_and_version_char(&self, arg: char) -> ClapResult<()>

fn use_long_help(&self) -> bool

fn _help(&self, use_long: bool) -> Error

fn _version(&self, use_long: bool) -> Error

fn parse_long_arg(
&mut self,
matcher: &mut ArgMatcher<'a>,
full_arg: &OsStr
) -> ClapResult<ParseResult<'a>>

fn parse_short_arg(
&mut self,
matcher: &mut ArgMatcher<'a>,
full_arg: &OsStr
) -> ClapResult<ParseResult<'a>>

fn parse_opt(
&self,
val: Option<&OsStr>,
opt: &OptBuilder<'a, 'b>,
had_eq: bool,
matcher: &mut ArgMatcher<'a>
) -> ClapResult<ParseResult<'a>>

fn add_val_to_arg<A>(
&self,
arg: &A,
val: &OsStr,
matcher: &mut ArgMatcher<'a>
) -> ClapResult<ParseResult<'a>> where
A: AnyArg<'a, 'b> + Display,

fn add_single_val_to_arg<A>(
&self,
arg: &A,
v: &OsStr,
matcher: &mut ArgMatcher<'a>
) -> ClapResult<ParseResult<'a>> where
A: AnyArg<'a, 'b> + Display,

fn parse_flag(
&self,
flag: &FlagBuilder<'a, 'b>,
matcher: &mut ArgMatcher<'a>
) -> ClapResult<ParseResult<'a>>

fn did_you_mean_error(
&self,
arg: &str,
matcher: &mut ArgMatcher<'a>
) -> ClapResult<()>

fn print_version<W: Write>(&self, w: &mut W, use_long: bool) -> ClapResult<()>

pub fn write_version<W: Write>(&self, w: &mut W, use_long: bool) -> Result<()>

pub fn print_help(&self) -> ClapResult<()>

pub fn write_help<W: Write>(&self, w: &mut W) -> ClapResult<()>

pub fn write_long_help<W: Write>(&self, w: &mut W) -> ClapResult<()>

pub fn write_help_err<W: Write>(&self, w: &mut W) -> ClapResult<()>

pub fn add_defaults(&mut self, matcher: &mut ArgMatcher<'a>) -> ClapResult<()>

pub fn add_env(&mut self, matcher: &mut ArgMatcher<'a>) -> ClapResult<()>

pub fn flags(&self) -> Iter<FlagBuilder<'a, 'b>>

pub fn opts(&self) -> Iter<OptBuilder<'a, 'b>>

pub fn positionals(&self) -> Values<'a, usize, PosBuilder<'a, 'b>>

pub fn subcommands(&self) -> Iter<App>

pub fn color(&self) -> ColorWhen

pub fn find_any_arg(&self, name: &str) -> Option<&AnyArg<'a, 'b>>

fn is_bin_name(&self, value: &str) -> bool

Check is a given string matches the binary name for this parser

fn is_alias(&self, value: &str) -> bool

Check is a given string is an alias for this parser

pub fn find_subcommand(&'b self, sc: &str) -> Option<&'b App<'a, 'b>>

fn contains_long(&self, l: &str) -> bool

fn contains_short(&self, s: char) -> bool