Struct clap::app::settings::Flags

struct Flags {
    bits: u64,
}

Fields
bits: u64
Methods

impl Flags

pub const SC_NEGATE_REQS: Flags

SC_NEGATE_REQS: Flags = Flags{bits: 1,}

pub const SC_REQUIRED: Flags

SC_REQUIRED: Flags = Flags{bits: 1 << 1,}

pub const A_REQUIRED_ELSE_HELP: Flags

A_REQUIRED_ELSE_HELP: Flags = Flags{bits: 1 << 2,}

pub const GLOBAL_VERSION: Flags

GLOBAL_VERSION: Flags = Flags{bits: 1 << 3,}

pub const VERSIONLESS_SC: Flags

VERSIONLESS_SC: Flags = Flags{bits: 1 << 4,}

pub const UNIFIED_HELP: Flags

UNIFIED_HELP: Flags = Flags{bits: 1 << 5,}

pub const WAIT_ON_ERROR: Flags

WAIT_ON_ERROR: Flags = Flags{bits: 1 << 6,}

pub const SC_REQUIRED_ELSE_HELP: Flags

SC_REQUIRED_ELSE_HELP: Flags = Flags{bits: 1 << 7,}

pub const NEEDS_LONG_HELP: Flags

NEEDS_LONG_HELP: Flags = Flags{bits: 1 << 8,}

pub const NEEDS_LONG_VERSION: Flags

NEEDS_LONG_VERSION: Flags = Flags{bits: 1 << 9,}

pub const NEEDS_SC_HELP: Flags

NEEDS_SC_HELP: Flags = Flags{bits: 1 << 10,}

pub const DISABLE_VERSION: Flags

DISABLE_VERSION: Flags = Flags{bits: 1 << 11,}

pub const HIDDEN: Flags

HIDDEN: Flags = Flags{bits: 1 << 12,}

pub const TRAILING_VARARG: Flags

TRAILING_VARARG: Flags = Flags{bits: 1 << 13,}

pub const NO_BIN_NAME: Flags

NO_BIN_NAME: Flags = Flags{bits: 1 << 14,}

pub const ALLOW_UNK_SC: Flags

ALLOW_UNK_SC: Flags = Flags{bits: 1 << 15,}

pub const UTF8_STRICT: Flags

UTF8_STRICT: Flags = Flags{bits: 1 << 16,}

pub const UTF8_NONE: Flags

UTF8_NONE: Flags = Flags{bits: 1 << 17,}

pub const LEADING_HYPHEN: Flags

LEADING_HYPHEN: Flags = Flags{bits: 1 << 18,}

pub const NO_POS_VALUES: Flags

NO_POS_VALUES: Flags = Flags{bits: 1 << 19,}

pub const NEXT_LINE_HELP: Flags

NEXT_LINE_HELP: Flags = Flags{bits: 1 << 20,}

pub const DERIVE_DISP_ORDER: Flags

DERIVE_DISP_ORDER: Flags = Flags{bits: 1 << 21,}

pub const COLORED_HELP: Flags

COLORED_HELP: Flags = Flags{bits: 1 << 22,}

pub const COLOR_ALWAYS: Flags

COLOR_ALWAYS: Flags = Flags{bits: 1 << 23,}

pub const COLOR_AUTO: Flags

COLOR_AUTO: Flags = Flags{bits: 1 << 24,}

pub const COLOR_NEVER: Flags

COLOR_NEVER: Flags = Flags{bits: 1 << 25,}

pub const DONT_DELIM_TRAIL: Flags

DONT_DELIM_TRAIL: Flags = Flags{bits: 1 << 26,}

pub const ALLOW_NEG_NUMS: Flags

ALLOW_NEG_NUMS: Flags = Flags{bits: 1 << 27,}

pub const LOW_INDEX_MUL_POS: Flags

LOW_INDEX_MUL_POS: Flags = Flags{bits: 1 << 28,}

pub const DISABLE_HELP_SC: Flags

DISABLE_HELP_SC: Flags = Flags{bits: 1 << 29,}

pub const DONT_COLLAPSE_ARGS: Flags

DONT_COLLAPSE_ARGS: Flags = Flags{bits: 1 << 30,}

pub const ARGS_NEGATE_SCS: Flags

ARGS_NEGATE_SCS: Flags = Flags{bits: 1 << 31,}

pub const PROPAGATE_VALS_DOWN: Flags

PROPAGATE_VALS_DOWN: Flags = Flags{bits: 1 << 32,}

pub const ALLOW_MISSING_POS: Flags

ALLOW_MISSING_POS: Flags = Flags{bits: 1 << 33,}

pub const TRAILING_VALUES: Flags

TRAILING_VALUES: Flags = Flags{bits: 1 << 34,}

pub const VALID_NEG_NUM_FOUND: Flags

VALID_NEG_NUM_FOUND: Flags = Flags{bits: 1 << 35,}

pub const PROPAGATED: Flags

PROPAGATED: Flags = Flags{bits: 1 << 36,}

pub const VALID_ARG_FOUND: Flags

VALID_ARG_FOUND: Flags = Flags{bits: 1 << 37,}

pub const INFER_SUBCOMMANDS: Flags

INFER_SUBCOMMANDS: Flags = Flags{bits: 1 << 38,}

pub const CONTAINS_LAST: Flags

CONTAINS_LAST: Flags = Flags{bits: 1 << 39,}

pub fn empty() -> Flags

Returns an empty set of flags.

pub fn all() -> Flags

Returns the set containing all flags.

pub fn bits(&self) -> u64

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u64) -> Option<Flags>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub fn from_bits_truncate(bits: u64) -> Flags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub fn intersects(&self, other: Flags) -> bool

Returns true if there are flags common to both self and other.

pub fn contains(&self, other: Flags) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: Flags)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Flags)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Flags)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Flags, value: bool)

Inserts or removes the specified flags depending on the passed value.