Struct clap::builder::ValueRange

pub struct ValueRange { /* private fields */ }

Values per occurrence for an argument

Implementations

impl ValueRange

pub const EMPTY: ValueRange = _

Nor argument values, or a flag

pub const SINGLE: ValueRange = _

A single argument value, the most common case for options

pub fn new(range: impl Into<ValueRange>) -> ValueRange

Create a range

Panics

If the end is less than the start (debug builds)

Examples

let range = ValueRange::new(5);
let range = ValueRange::new(5..10);
let range = ValueRange::new(5..=10);
let range = ValueRange::new(5..);
let range = ValueRange::new(..10);
let range = ValueRange::new(..=10);

While this will panic:

let range = ValueRange::new(10..5);  // Panics!

pub fn min_values(&self) -> usize

Fewest number of values the argument accepts

pub fn max_values(&self) -> usize

Most number of values the argument accepts

pub fn takes_values(&self) -> bool

Report whether the argument takes any values (ie is a flag)

Examples

let range = ValueRange::new(5);
assert!(range.takes_values());

let range = ValueRange::new(0);
assert!(!range.takes_values());