Enum clap::Format

pub enum Format<T> {
    Error(T),
    Warning(T),
    Good(T),
    None(T),
}

Defines styles for different types of error messages. Defaults to Error=Red, Warning=Yellow, and Good=Green

Variants
Error(T)

Defines the style used for errors, defaults to Red

Warning(T)

Defines the style used for warnings, defaults to Yellow

Good(T)

Defines the style used for good values, defaults to Green

None(T)

Defines no formatting style

Methods

impl<T: AsRef<str>> Format<T>

fn format(&self) -> ANSIString