Struct clap::error::Error

pub struct Error {
    pub kind: ErrorKind,
    pub info: Vec<String>,
    /* private fields */
}

Command Line Argument Parser Error

See Command::error to create an error.

Fields
kind: ErrorKind

Deprecated, replaced with Error::kind()

info: Vec<String>

Deprecated, replaced with Error::context()

Implementations

impl Error

pub fn raw(kind: ErrorKind, message: impl Display) -> Self

Create an unformatted error

This is for you need to pass the error up to a place that has access to the Command at which point you can call Error::format.

Prefer Command::error for generating errors.

pub fn format(self, cmd: &mut Command<'_>) -> Self

Format the existing message with the Command’s context

pub fn kind(&self) -> ErrorKind

Type of error for programmatic processing

pub fn context(&self) -> impl Iterator<Item = (ContextKind, &ContextValue)>

Additional information to further qualify the error

pub fn use_stderr(&self) -> bool

Should the message be written to stdout or not?

pub fn exit(&self) -> !

Prints the error and exits.

Depending on the error kind, this either prints to stderr and exits with a status of 2 or prints to stdout and exits with a status of 0.

pub fn print(&self) -> Result<()>

Prints formatted and colored error to stdout or stderr according to its error kind

Example

use clap::Command;

match Command::new("Command").try_get_matches() {
    Ok(matches) => {
        // do_something
    },
    Err(err) => {
        err.print().expect("Error writing Error");
        // do_something
    },
};