Macro clap::_clap_count_exprs

macro_rules! _clap_count_exprs {
    () => { ... };
    ($e:expr) => { ... };
    ($e:expr, $($es:expr),+) => { ... };
}

Counts the number of comma-delimited expressions passed to it. The result is a compile-time evaluable expression, suitable for use as a static array size, or the value of a const.

Examples

const COUNT: usize = _clap_count_exprs!(a, 5+1, "hi there!".into_string());
assert_eq!(COUNT, 3);