Struct clap::Error

pub struct Error {
    pub message: String,
    pub kind: ErrorKind,
    pub info: Option<Vec<String>>,
}

Command Line Argument Parser Error

Fields
message: String

Formated error message

kind: ErrorKind

The type of error

info: Option<Vec<String>>

Any additional information passed along, such as the argument name that caused the error

Methods

impl Error

pub fn use_stderr(&self) -> bool

Should the message be written to stdout or not

pub fn exit(&self) -> !

Prints the error to stderr and exits with a status of 1

pub fn write_to<W: Write>(&self, w: &mut W) -> Result<()>

pub fn argument_conflict<'a, 'b, O, U>(
arg: &AnyArg,
other: Option<O>,
usage: U,
color: ColorWhen
) -> Self where
O: Into<String>,
U: Display,

pub fn empty_value<'a, 'b, U>(arg: &AnyArg, usage: U, color: ColorWhen) -> Self where
U: Display,

pub fn invalid_value<'a, 'b, B, G, U>(
bad_val: B,
good_vals: &[G],
arg: &AnyArg,
usage: U,
color: ColorWhen
) -> Self where
B: AsRef<str>,
G: AsRef<str> + Display,
U: Display,

pub fn invalid_subcommand<S, D, N, U>(
subcmd: S,
did_you_mean: D,
name: N,
usage: U,
color: ColorWhen
) -> Self where
S: Into<String>,
D: AsRef<str> + Display,
N: Display,
U: Display,

pub fn unrecognized_subcommand<S, N>(
subcmd: S,
name: N,
color: ColorWhen
) -> Self where
S: Into<String>,
N: Display,

pub fn missing_required_argument<R, U>(
required: R,
usage: U,
color: ColorWhen
) -> Self where
R: Display,
U: Display,

pub fn missing_subcommand<N, U>(name: N, usage: U, color: ColorWhen) -> Self where
N: AsRef<str> + Display,
U: Display,

pub fn invalid_utf8<U>(usage: U, color: ColorWhen) -> Self where
U: Display,

pub fn too_many_values<'a, 'b, V, U>(
val: V,
arg: &AnyArg,
usage: U,
color: ColorWhen
) -> Self where
V: AsRef<str> + Display + ToOwned,
U: Display,

pub fn too_few_values<'a, 'b, U>(
arg: &AnyArg,
min_vals: u64,
curr_vals: usize,
usage: U,
color: ColorWhen
) -> Self where
U: Display,

pub fn value_validation<'a, 'b>(
arg: Option<&AnyArg>,
err: String,
color: ColorWhen
) -> Self

pub fn value_validation_auto(err: String) -> Self

pub fn wrong_number_of_values<'a, 'b, S, U>(
arg: &AnyArg,
num_vals: u64,
curr_vals: usize,
suffix: S,
usage: U,
color: ColorWhen
) -> Self where
S: Display,
U: Display,

pub fn unexpected_multiple_usage<'a, 'b, U>(
arg: &AnyArg,
usage: U,
color: ColorWhen
) -> Self where
U: Display,

pub fn unknown_argument<A, U>(
arg: A,
did_you_mean: &str,
usage: U,
color: ColorWhen
) -> Self where
A: Into<String>,
U: Display,

pub fn io_error(e: &Error, color: ColorWhen) -> Self

pub fn argument_not_found_auto<A>(arg: A) -> Self where
A: Into<String>,

pub fn with_description(description: &str, kind: ErrorKind) -> Self

Create an error with a custom description.

This can be used in combination with Error::exit to exit your program with a custom error message.