Trait clap::builder::TypedValueParser

pub trait TypedValueParser: Clone + Send + Sync + 'static {
    type Value;

    fn parse_ref(
        &self, 
        cmd: &Command<'_>, 
        arg: Option<&Arg<'_>>, 
        value: &OsStr
    ) -> Result<Self::Value, Error>;

    fn parse(
        &self, 
        cmd: &Command<'_>, 
        arg: Option<&Arg<'_>>, 
        value: OsString
    ) -> Result<Self::Value, Error> { ... }
    fn possible_values(
        &self
    ) -> Option<Box<dyn Iterator<Item = PossibleValue<'static>>>> { ... }
}

Parse/validate argument values

Required Associated Types

type Value

Argument’s value type

Required Methods

fn parse_ref(
&self,
cmd: &Command<'_>,
arg: Option<&Arg<'_>>,
value: &OsStr
) -> Result<Self::Value, Error>

Parse the argument value

When arg is None, an external subcommand value is being parsed.

Provided Methods

fn parse(
&self,
cmd: &Command<'_>,
arg: Option<&Arg<'_>>,
value: OsString
) -> Result<Self::Value, Error>

Parse the argument value

When arg is None, an external subcommand value is being parsed.

fn possible_values(
&self
) -> Option<Box<dyn Iterator<Item = PossibleValue<'static>>>>

Reflect on enumerated value properties

Error checking should not be done with this; it is mostly targeted at user-facing applications like errors and completion.