Enum clap::ArgSettings

#[non_exhaustive]
pub enum ArgSettings {

    Required,
    MultipleValues,
    MultipleOccurrences,
    Multiple,
    ForbidEmptyValues,
    Global,
    Hidden,
    TakesValue,
    UseValueDelimiter,
    NextLineHelp,
    RequireDelimiter,
    HidePossibleValues,
    AllowHyphenValues,
    AllowLeadingHyphen,
    RequireEquals,
    Last,
    HideDefaultValue,
    IgnoreCase,
    CaseInsensitive,
    HiddenShortHelp,
    HiddenLongHelp,
    AllowInvalidUtf8,
    Exclusive,

}

Various settings that apply to arguments and may be set, unset, and checked via getter/setter methods Arg::setting, Arg::unset_setting, and Arg::is_set. This is what the Arg methods which accept a bool use internally.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Required

Specifies that an arg must be used

MultipleValues

Allows an arg to accept multiple values

MultipleOccurrences

Allows an arg to appear multiple times

Multiple

👎 Deprecated since 3.0.0:
Split into ArgSettings::MultipleOccurrences (most likely what you want) and ArgSettings::MultipleValues

Deprecated, see ArgSettings::MultipleOccurrences (most likely what you want) and ArgSettings::MultipleValues

ForbidEmptyValues

Forbids an arg from accepting empty values such as ""

Global

Sets an arg to be global (i.e. exist in all subcommands)

Hidden

Hides an arg from the help message

TakesValue

Allows an argument to take a value (such as --option value)

UseValueDelimiter

Enables a delimiter to break up arguments --option val1,val2,val3 becomes three values (val1, val2, and val3) instead of the default one (val1,val2,val3)

NextLineHelp

Tells an arg to display it’s help on the line below the arg itself in the help message

RequireDelimiter

Says that arg must use a delimiter to separate values

HidePossibleValues

Hides the possible values from the help message

AllowHyphenValues

Allows values that start with a hyphen

AllowLeadingHyphen

👎 Deprecated since 3.0.0:
Replaced with ArgSettings::AllowHyphenValues

Deprecated, replaced with ArgSettings::AllowHyphenValues

RequireEquals

Requires that an equals be used to provide a value to an option such as --option=value

Last

Says that a positional arg will be the last positional, and requires -- to be accessed. It can also be accessed early (i.e. before other positionals) by providing --

HideDefaultValue

Hides the default value from the help message

IgnoreCase

Possible values become case insensitive

CaseInsensitive

👎 Deprecated since 3.0.0:
Replaced with ArgSettings::IgnoreCase

Deprecated, replaced with ArgSettings::IgnoreCase

HiddenShortHelp

The argument should not be shown in short help text

HiddenLongHelp

The argument should not be shown in long help text

AllowInvalidUtf8

Specifies that option values that are invalid UTF-8 should not be treated as an error.

Exclusive

Specifies that option should exist on its own. Having any other arguments present at runtime is an error.