Enum clap::ArgSettings

pub enum ArgSettings {
    Required,
    Multiple,
    EmptyValues,
    Global,
    Hidden,
    TakesValue,
    UseValueDelimiter,
}

Various settings that apply to arguments and may be set, unset, and checked via getter/setter methods Arg::set, Arg::unset, and Arg::is_set

Variants

Required

The argument must be used

Multiple

The argument may be used multiple times such as --flag --flag

EmptyValues

The argument allows empty values such as --option ""

Global

The argument should be propagated down through all child subcommands

Hidden

The argument should not be shown in help text

TakesValue

The argument accepts a value, such as --option <value>

UseValueDelimiter

Determines if the argument allows values to be grouped via a delimter