Enum clap::ArgSettings

pub enum ArgSettings {
    Required,
    Multiple,
    EmptyValues,
    Global,
    Hidden,
    TakesValue,
    UseValueDelimiter,
    NextLineHelp,
    RequireDelimiter,
    HidePossibleValues,
    AllowLeadingHyphen,
    RequireEquals,
    Last,
    HideDefaultValue,
    CaseInsensitive,
    HideEnvValues,
    HiddenShortHelp,
    HiddenLongHelp,
    // some variants omitted
}

Various settings that apply to arguments and may be set, unset, and checked via getter/setter methods Arg::set, Arg::unset, and Arg::is_set

Variants

Required

The argument must be used

Multiple

The argument may be used multiple times such as --flag --flag

EmptyValues

The argument allows empty values such as --option ""

Global

The argument should be propagated down through all child SubCommands

Hidden

The argument should not be shown in help text

TakesValue

The argument accepts a value, such as --option <value>

UseValueDelimiter

Determines if the argument allows values to be grouped via a delimiter

NextLineHelp

Prints the help text on the line after the argument

RequireDelimiter

Requires the use of a value delimiter for all multiple values

HidePossibleValues

Hides the possible values from the help string

AllowLeadingHyphen

Allows vals that start with a '-'

RequireEquals

Require options use --option=val syntax

Last

Specifies that the arg is the last positional argument and may be accessed early via -- syntax

HideDefaultValue

Hides the default value from the help string

CaseInsensitive

Makes Arg::possible_values case insensitive

HideEnvValues

Hides ENV values in the help message

HiddenShortHelp

The argument should not be shown in short help text

HiddenLongHelp

The argument should not be shown in long help text