Module clap::parser

Command line argument parser

Structs

ArgMatches

Container for parse results.

Indices

Iterate over indices for where an argument appeared when parsing, via ArgMatches::indices_of

OsValues

Deprecated, replaced with ArgMatches::get_many()

RawValues

Iterate over raw argument values via ArgMatches::get_raw.

Values

Deprecated, replaced with ArgMatches::get_many()

ValuesRef

Iterate over multiple values for an argument via ArgMatches::get_many.

Enums

MatchesError

Violation of ArgMatches assumptions

ValueSource

Origin of the argument’s value