Struct clap::App

pub struct App {
    pub name: &'static str,
    pub author: Option<&'static str>,
    pub version: Option<&'static str>,
    pub about: Option<&'static str>,
    // some fields omitted
}

Used to create a representation of the program and all possible command line arguments for parsing at runtime.

Stores a list of all posisble arguments, as well as information displayed to the user such as help and versioning information.

Example:

 let myprog = App::new("myprog")
                   .author("Me, me@mail.com")
                      .version("1.0.2")
                   .about("Explains in brief what the program does")
                   .arg(
                            Arg::new("in_file").index(1)
                        // Add other possible command line argument options here...
                    )
                   .get_matches();

 // Your pogram logic starts here...

Fields

name

The name displayed to the user when showing version and help/usage information

author

A string of author(s) if desired. Displayed when showing help/usage information

version

The version displayed to the user

about

A brief explaination of the program that gets displayed to the user when shown help/usage information

Methods

impl App

fn new(n: &'static str) -> App

Creates a new instance of an application requiring a name (such as the binary). Will be displayed to the user when they print version or help and usage information.

Example:

let prog = App::new("myprog")

fn author(&mut self, a: &'static str) -> &mut App

Sets a string of author(s)

Example:

.author("Kevin <kbknapp@gmail.com>")

fn about(&mut self, a: &'static str) -> &mut App

Sets a string briefly describing what the program does

Example:

.about("Does really amazing things to great people")

fn version(&mut self, v: &'static str) -> &mut App

Sets a string of the version number

Example:

.version("v0.1.24")

fn arg(&mut self, a: &Arg) -> &mut App

Adds an argument to the list of valid possibilties

Example:

 .arg(Arg::new("config")
                .short("c")
            // Additional argument configuration goes here...
 )

fn args(&mut self, args: Vec<&Arg>) -> &mut App

Adds arguments to the list of valid possibilties

Example:

 .args( vec![Arg::new("config").short("c"),
                Arg::new("debug").short("d")])

fn get_matches(&mut self) -> ArgMatches