Struct clap::ArgMatches

pub struct ArgMatches {
    pub name: &'static str,
    pub flags: HashMap<&'static str, FlagArg>,
    pub opts: HashMap<&'static str, OptArg>,
    pub positionals: HashMap<&'static str, PosArg>,
}

Used to get information about the arguments that where supplied to the program at runtime.

Fields of ArgMatches aren't designed to be used directly, only the methods in order to query information.

 let matches = App::new("MyApp")
// adding of arguments and configuration goes here...
                    .get_matches();
   // if you had an argument named "output" that takes a value 
   if let Some(o) = matches.value_of("output") {
       println!("Value for output: {}", o);
   }

   // Although not advised, if you have a required argument
   // you can call .unwrap() because the program will exit long before
   // here at noticing the user didn't supply a required argument...
   // use at your own risk ;)
   println!("Config file: {}", matches.value_of("config").unwrap());

   // You can check the present of an argument
   if matches.is_present("debug") {
       // Checking if "debug" was present was necessary,
       // as occurrences returns 0 if a flag isn't found
       // but we can check how many times "debug" was found
       // if we allow multiple (if multiple isn't allowed it always be 1 or 0)
       if matches.occurrences_of("debug") > 1 {
           println!("Debug mode is REALLY on");
       } else {
           println!("Debug mode kind of on");
       }
}

Fields

name

flags

opts

positionals

Methods

impl ArgMatches

fn new(app: &App) -> ArgMatches

Creates a new instance of ArgMatches. This ins't called directly, but through the .get_matches() method of App

Example:

let matches = App::new("myprog").get_matches();

fn value_of(&self, name: &'static str) -> Option<&String>

Gets the value of a specific option or positional argument (i.e. an argument that takes an additional value at runtime). If the option wasn't present at runtime it returns None

Example:

if let Some(o) = matches.value_of("output") {
       println!("Value for output: {}", o);
}

fn is_present(&self, name: &'static str) -> bool

Checks if a flag was argument was supplied at runtime. DOES NOT work for option or positional arguments (use .value_of() instead)

Example:

if matches.is_present("output") {
       println!("The output argument was used!");
}

fn occurrences_of(&self, name: &'static str) -> u8

Checks the number of occurrences of a flag at runtime.

This DOES NOT work for option or positional arguments (use .value_of() instead). If a flag isn't present it will return 0, if a flag doesn't allow multiple occurrences, it will return 1 no matter how many times it occurred (unless it wasn't prsent) at all.

Example:

if matches.occurrences_of("debug") > 1 {
    println!("Debug mode is REALLY on");
} else {
    println!("Debug mode kind of on");
}