Struct clap::Error

pub struct Error {
    pub message: String,
    pub kind: ErrorKind,
    pub info: Option<Vec<String>>,
}

Command Line Argument Parser Error

Fields
message: String

Formatted error message

kind: ErrorKind

The type of error

info: Option<Vec<String>>

Any additional information passed along, such as the argument name that caused the error

Implementations

impl Error

pub fn use_stderr(&self) -> bool

Should the message be written to stdout or not

pub fn exit(&self) -> !

Prints the error message and exits. If Error::use_stderr evaluates to true, the message will be written to stderr and exits with a status of 1. Otherwise, stdout is used with a status of 0.

pub fn with_description(description: &str, kind: ErrorKind) -> Self

Create an error with a custom description.

This can be used in combination with Error::exit to exit your program with a custom error message.