Struct clap::SubCommand

pub struct SubCommand<'a> { /* fields omitted */ }

The abstract representation of a command line subcommand.

This struct describes all the valid options of the subcommand for the program. Subcommands are essentially "sub-Apps" and contain all the same possibilities (such as their own arguments, subcommands, and settings).

Examples

App::new("myprog")
    .subcommand(
        SubCommand::with_name("config")
            .about("Used for configuration")
            .arg(Arg::with_name("config_file")
                .help("The configuration file to use")
                .index(1)))

Methods

impl<'a> SubCommand<'a>

fn with_name<'b>(name: &str) -> App<'a, 'b>

Creates a new instance of a subcommand requiring a name. The name will be displayed to the user when they print version or help and usage information.

Examples

App::new("myprog")
    .subcommand(
        SubCommand::with_name("config"))