Trait clap::Subcommand

pub trait Subcommand: FromArgMatches + Sized {
    fn augment_subcommands(app: App<'_>) -> App<'_>;

    fn augment_subcommands_for_update(app: App<'_>) -> App<'_>;

    fn has_subcommand(name: &str) -> bool;
}

Parse a sub-command into a user-defined enum.

Implementing this trait lets a parent container delegate subcommand behavior to Self. with:

#[clap(subcommand)] field: SubCmd: Attribute can be used with either struct fields or enum variants that impl Subcommand.

#[clap(flatten)] Variant(SubCmd): Attribute can only be used with enum variants that impl Subcommand.

See the derive reference for attributes and best practices.

NOTE: Deriving requires the derive feature flag

Example

#[derive(clap::Parser)]
struct Args {
   #[clap(subcommand)]
   action: Action,
}

#[derive(clap::Subcommand)]
enum Action {
   Add,
   Remove,
}

Required methods

fn augment_subcommands(app: App<'_>) -> App<'_>

Append to App so it can instantiate Self.

See also IntoApp.

fn augment_subcommands_for_update(app: App<'_>) -> App<'_>

Append to App so it can update self.

This is used to implement #[clap(flatten)]

See also IntoApp.

fn has_subcommand(name: &str) -> bool

Test whether Self can parse a specific subcommand